Eliminate the double write barrier on SIMPLE-VECTORs.
Scavenging looks for old->new pointers between the start of a card
and the start of the first object on the card, then searches
backward for a SIMPLE-VECTOR with that pointer to scavenge. This
slows down scavenging, but reduces mutator overhead and eliminates
the false-sharing that the double write barrier had.